NumericLinearFit now takes parameter and derivative types as template parameters Added header with some common typedefs Added automatic generation of expression template engine for any iterable class Added example program performing a simultaneous linear fit on 2 data types

+++ include/distribution.h
@@ -86,7 +86,7 @@ public:
   const DataType & sample(const int idx) const{ return _data[idx]; }
   DataType & sample(const int idx){ return _data[idx]; }
 
-  DataType mean() const{ return threadedSum(_data)/_data.size(); }
+  DataType mean() const{ return threadedSum(_data)/double(_data.size()); }
 
   DataType variance() const{
     const int N = _data.size();
@@ -177,7 +177,7 @@ public:
       inline DataType operator()(const int i) const{ DataType tmp = data[i] - avg; return tmp * tmp;  }
     };
     Op op(this->mean(),this->_data);
-    return sqrt( threadedSum(op) * double(N-1)/N );
+    return sqrt( threadedSum(op) * (double(N-1)/N) );
   }
   
   jackknifeDistribution(): distribution<DataType>(){}
